% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucminf.R
\name{print.ucminf}
\alias{print.ucminf}
\title{Print method for ucminf objects}
\usage{
\method{print}{ucminf}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{A \code{ucminf} object.}

\item{digits}{Number of significant digits to print.}

\item{...}{Unused.}
}
\description{
Print method for ucminf objects
}
