% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucminf_xptr.R
\name{ucminf_xptr}
\alias{ucminf_xptr}
\title{Optimize using a compiled C++ objective (XPtr interface)}
\usage{
ucminf_xptr(par, xptr, control = list(), hessian = 0)
}
\arguments{
\item{par}{Numeric starting vector.}

\item{xptr}{An \code{externalptr} created by wrapping a
\code{ucminf::ObjFun*} in \code{Rcpp::XPtr<ucminf::ObjFun>}.}

\item{control}{A named list of control parameters (see \code{\link{ucminf}}),
or a \code{\link{ucminf_control}} object.}

\item{hessian}{Integer: 0 = none, 2 = inv-Hessian, 3 = both.}
}
\value{
A list of class \code{"ucminf"} (same structure as \code{\link{ucminf}}).
}
\description{
Calls the UCMINF optimizer with a compiled C++ objective function
that has been wrapped in an \code{Rcpp::XPtr<ucminf::ObjFun>}.
This path bypasses the R interpreter on every function evaluation,
giving maximum performance for non-trivial objectives.
}
\seealso{
\code{\link{ucminf}}, \code{\link{ucminf_control}}
}
